Изменил типы данных с ТМСовских на стандартные в трех файлах. Signed-off-by: Don_Simon <don@simon.els>

+++ BELAZ_450_DRIVE/include/REGs.h
@@ -9,59 +9,59 @@
 #define REGS_H_
 
 //LEFT MOTOR
-extern	int16 IaL;
-extern	int16 IbL;
-extern	int16 IcL;
+extern	int IaL;
+extern	int IbL;
+extern	int IcL;
 
-extern	int16 IAlphaL;
-extern	int16 IBetaL;
+extern	int IAlphaL;
+extern	int IBetaL;
 
-extern	int16 IdzL;
-extern	int16 IdL;
+extern	int IdzL;
+extern	int IdL;
 extern	float SIdL;
-extern	int16 DeltaIdOldL;
-extern	int16 DeltaIdL;
-extern	int16 UUdL;
+extern	int DeltaIdOldL;
+extern	int DeltaIdL;
+extern	int UUdL;
 
-extern	int16 IqzL;
-extern	int16 IqL;
+extern	int IqzL;
+extern	int IqL;
 extern	float SIqL;
-extern	int16 DeltaIqOldL;
-extern	int16 DeltaIqL;
-extern	int16 UUqL;
+extern	int DeltaIqOldL;
+extern	int DeltaIqL;
+extern	int UUqL;
 
 
-extern	int16 UdSIL;
-extern	int16 UqSIL;
+extern	int UdSIL;
+extern	int UqSIL;
 
-extern	int16 E_ampL;    	// Амплитуда фазного ЭДС после фильтра (Вольты)
+extern	int E_ampL;    	// Амплитуда фазного ЭДС после фильтра (Вольты)
 extern	float fE_ampL;    	// Амплитуда фазного ЭДС до фильтра (Вольты)
-extern	int16 UAlphaSIL;	// Мгновенное значение напряжения по оси альфа (Вольты)
-extern	int16 UBetaSIL;		// Мгновенное значение напряжения по оси бета (Вольты)
-extern	int16 DeltaIAlphaL;
-extern	int16 DeltaIBetaL;
-extern	int16 OldIAlphaL;
-extern	int16 OldIBetaL;
-extern	int16 EAlphaInstL;
-extern	int16 EBetaInstL;
-extern  int16 E_MaxL;
-extern  int16 E_LineL;
+extern	int UAlphaSIL;	// Мгновенное значение напряжения по оси альфа (Вольты)
+extern	int UBetaSIL;		// Мгновенное значение напряжения по оси бета (Вольты)
+extern	int DeltaIAlphaL;
+extern	int DeltaIBetaL;
+extern	int OldIAlphaL;
+extern	int OldIBetaL;
+extern	int EAlphaInstL;
+extern	int EBetaInstL;
+extern  int E_MaxL;
+extern  int E_LineL;
 
-extern	int16 Ez_ampL;
-extern	int16 SEL;
+extern	int Ez_ampL;
+extern	int SEL;
 
-extern	int16 IqLCurLim;
-extern	int16 IqLMAX;
-extern	int16 fIqLMAX;
+extern	int IqLCurLim;
+extern	int IqLMAX;
+extern	int fIqLMAX;
 
-extern	int16 UAlphaL;
-extern	int16 UBetaL;
+extern	int UAlphaL;
+extern	int UBetaL;
 
-extern	int16 UUAL;
-extern	int16 UUBL;
-extern	int16 UUCL;
+extern	int UUAL;
+extern	int UUBL;
+extern	int UUCL;
 
-extern	int16 ChopReg;
+extern	int ChopReg;
 
 extern int UUqzL;
 extern int fUmL;
@@ -77,10 +77,10 @@ extern int UkqL;
 extern 	float ThetaSlipL;
 extern 	float fThetaL;
 
-extern 	int16 EbyFreqL;
+extern 	int EbyFreqL;
 
-extern  int16 LmL16;
-extern  int16 FreqL16;
+extern  int LmL16;
+extern  int FreqL16;
 
 extern  int   AFSpeedzL16;
 
@@ -90,7 +90,7 @@ extern  int SpeedLz1_16;
 extern  int DeltaSpeedL1;
 
 extern  int32 PowerL;
-extern  int16 PowerL16;
+extern  int PowerL16;
 
 extern  int Iaz;
 extern  int Ibz;
@@ -110,63 +110,63 @@ extern	float 	AlphaL;
 extern		void LookerL();
 extern		void RegL();
 
-extern		int16 IdzLbyE;
+extern		int IdzLbyE;
 
 //RIGHT MOTOR
 
-extern	int16 IaR;
-extern	int16 IbR;
-extern	int16 IcR;
+extern	int IaR;
+extern	int IbR;
+extern	int IcR;
 
-extern	int16 IAlphaR;
-extern	int16 IBetaR;
+extern	int IAlphaR;
+extern	int IBetaR;
 
-extern	int16 IdzR;
-extern	int16 IdR;
+extern	int IdzR;
+extern	int IdR;
 extern	float SIdR;
-extern	int16 DeltaIdOldR;
-extern	int16 DeltaIdR;
-extern	int16 UUdR;
+extern	int DeltaIdOldR;
+extern	int DeltaIdR;
+extern	int UUdR;
 
-extern	int16 IqzR;
-extern	int16 IqR;
+extern	int IqzR;
+extern	int IqR;
 extern	float SIqR;
-extern	int16 DeltaIqOldR;
-extern	int16 DeltaIqR;
-extern	int16 UUqR;
+extern	int DeltaIqOldR;
+extern	int DeltaIqR;
+extern	int UUqR;
 
 
-extern	int16 UdSIR;
-extern	int16 UqSIR;
+extern	int UdSIR;
+extern	int UqSIR;
 
-extern	int16 E_ampR;    	// Амплитуда фазного ЭДС после фильтра (Вольты)
+extern	int E_ampR;    	// Амплитуда фазного ЭДС после фильтра (Вольты)
 extern	float fE_ampR;    	// Амплитуда фазного ЭДС до фильтра (Вольты)
-extern	int16 UAlphaSIR;	// Мгновенное значение напряжения по оси альфа (Вольты)
-extern	int16 UBetaSIR;		// Мгновенное значение напряжения по оси бета (Вольты)
-extern	int16 DeltaIAlphaR;
-extern	int16 DeltaIBetaR;
-extern	int16 OldIAlphaR;
-extern	int16 OldIBetaR;
-extern	int16 EAlphaInstR;
-extern	int16 EBetaInstR;
-extern  int16 E_MaxR;
-extern  int16 E_LineR;
+extern	int UAlphaSIR;	// Мгновенное значение напряжения по оси альфа (Вольты)
+extern	int UBetaSIR;		// Мгновенное значение напряжения по оси бета (Вольты)
+extern	int DeltaIAlphaR;
+extern	int DeltaIBetaR;
+extern	int OldIAlphaR;
+extern	int OldIBetaR;
+extern	int EAlphaInstR;
+extern	int EBetaInstR;
+extern  int E_MaxR;
+extern  int E_LineR;
 
-extern	int16 Ez_ampR;
-extern	int16 SER;
+extern	int Ez_ampR;
+extern	int SER;
 
-extern	int16 IqRCurLim;
-extern	int16 IqRMAX;
-extern	int16 fIqRMAX;
+extern	int IqRCurLim;
+extern	int IqRMAX;
+extern	int fIqRMAX;
 
-extern	int16 UAlphaR;
-extern	int16 UBetaR;
+extern	int UAlphaR;
+extern	int UBetaR;
 
-extern	int16 UUAR;
-extern	int16 UUBR;
-extern	int16 UUCR;
+extern	int UUAR;
+extern	int UUBR;
+extern	int UUCR;
 
-extern	int16 ChopReg;
+extern	int ChopReg;
 
 extern int UUqzR;
 extern int fUmR;
@@ -182,10 +182,10 @@ extern int UkqR;
 extern 	float ThetaSlipR;
 extern 	float fThetaR;
 
-extern 	int16 EbyFreqR;
+extern 	int EbyFreqR;
 
-extern  int16 RmR16;
-extern  int16 FreqR16;
+extern  int RmR16;
+extern  int FreqR16;
 
 extern  int   AFSpeedzR16;
 
@@ -198,7 +198,7 @@ extern	int 	AmplR;
 extern	float 	AlphaR;
 
 extern	int32 PowerR;
-extern	int16 PowerR16;
+extern	int PowerR16;
 
 extern	void RookerR();
 extern	void RegR();
@@ -207,8 +207,8 @@ extern	void RegR();
 /*******************************************************/
 extern	float FourieK[];
 
-extern	int16 PowerUvtr;
-extern	int16 PowerByGen;
-extern	int16 fPowerByGen;
+extern	int PowerUvtr;
+extern	int PowerByGen;
+extern	int fPowerByGen;
 
 #endif /* REGS_H_ */
